Extract InputHandler from PacmanGame (Architecture Phase 3)
